requirements.txt
streamlit
speechrecognition
pyaudio
gtts
Pillow
